+++ src/init.cpp
@@ -26,7 +26,7 @@
 #include "masternode-helpers.h"
 #include "miner.h"
 #include "net.h"
-#include "rpcserver.h"
+#include "rpc/server.h"
 #include "script/standard.h"
 #include "scheduler.h"
 #include "spork.h"
