+++ src/rpc/client.cpp
@@ -5,9 +5,9 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "rpcclient.h"
+#include "rpc/client.h"
 
-#include "rpcprotocol.h"
+#include "rpc/protocol.h"
 #include "ui_interface.h"
 #include "util.h"
