+++ src/rpc/misc.cpp
@@ -12,7 +12,7 @@
 #include "masternode-sync.h"
 #include "net.h"
 #include "netbase.h"
-#include "rpcserver.h"
+#include "rpc/server.h"
 #include "spork.h"
 #include "timedata.h"
 #include "util.h"
